pyproject.toml
[project]
name = "slack-assistant-bot-client"
version = "0.1.0"
description = "Slack Assistant Bot Client for sending messages"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "slack-sdk>=3.26.1",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.4",
    "pydantic-settings>=2.7.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
